+++ android/engine/src/main/java/org/smartregister/fhircore/engine/ui/login/LoginViewModel.kt
@@ -173,6 +173,8 @@ constructor(
   val loginViewConfiguration: LiveData<LoginViewConfiguration>
     get() = _loginViewConfiguration
 
+  lateinit var appLogoResourceFile: String
+
   fun loginUser() {
     viewModelScope.launch(dispatcher.io()) {
       if (accountAuthenticator.hasActiveSession()) {