+++ android/engine/src/test/java/org/smartregister/fhircore/engine/ui/userprofile/UserProfileViewModelTest.kt
@@ -122,7 +122,7 @@ class UserProfileViewModelTest : RobolectricTest() {
 
   @Test
   fun loadSelectedLanguage() {
-    every { sharedPreferencesHelper.read(SharedPreferencesHelper.LANG, any()) } returns "fr"
+    every { sharedPreferencesHelper.read(SharedPreferencesHelper.LANG, "en") } returns "fr"
 
     Assert.assertEquals("French", userProfileViewModel.loadSelectedLanguage())
     verify { sharedPreferencesHelper.read(SharedPreferencesHelper.LANG, "en") }