+++ android/quest/src/main/java/org/smartregister/fhircore/quest/ui/patient/details/QuestPatientDetailActivity.kt
@@ -26,6 +26,7 @@ import ca.uhn.fhir.context.FhirContext
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
 import org.hl7.fhir.r4.model.QuestionnaireResponse
+import org.hl7.fhir.r4.model.ResourceType
 import org.smartregister.fhircore.engine.configuration.ConfigurationRegistry
 import org.smartregister.fhircore.engine.configuration.view.ConfigurableComposableView
 import org.smartregister.fhircore.engine.configuration.view.RegisterViewConfiguration
@@ -36,11 +37,10 @@ import org.smartregister.fhircore.engine.ui.questionnaire.QuestionnaireActivity.
 import org.smartregister.fhircore.engine.ui.questionnaire.QuestionnaireConfig
 import org.smartregister.fhircore.engine.ui.questionnaire.QuestionnaireType
 import org.smartregister.fhircore.engine.ui.theme.AppTheme
-import org.smartregister.fhircore.engine.util.AssetUtil
 import org.smartregister.fhircore.engine.util.extension.getEncounterId
 import org.smartregister.fhircore.quest.R
+import org.smartregister.fhircore.quest.configuration.view.DataDetailsListViewConfiguration
 import org.smartregister.fhircore.quest.configuration.view.NavigationOption
-import org.smartregister.fhircore.quest.configuration.view.PatientDetailsViewConfiguration
 import org.smartregister.fhircore.quest.configuration.view.QuestionnaireNavigationAction
 import org.smartregister.fhircore.quest.configuration.view.ResultDetailsNavigationConfiguration
 import org.smartregister.fhircore.quest.configuration.view.TestDetailsNavigationAction
@@ -50,14 +50,13 @@ import org.smartregister.fhircore.quest.util.QuestConfigClassification
 
 @AndroidEntryPoint
 class QuestPatientDetailActivity :
-  BaseMultiLanguageActivity(), ConfigurableComposableView<PatientDetailsViewConfiguration> {
+  BaseMultiLanguageActivity(), ConfigurableComposableView<DataDetailsListViewConfiguration> {
 
-  private lateinit var profileConfig: QuestPatientDetailViewModel.ProfileConfig
   var patientResourcesList: ArrayList<String> = arrayListOf()
-  private lateinit var patientDetailConfig: PatientDetailsViewConfiguration
+  private lateinit var patientDetailConfig: DataDetailsListViewConfiguration
   private lateinit var patientId: String
 
-  val patientViewModel by viewModels<QuestPatientDetailViewModel>()
+  val patientViewModel by viewModels<ListDataDetailViewModel>()
 
   @Inject lateinit var configurationRegistry: ConfigurationRegistry
 
@@ -78,22 +77,16 @@ class QuestPatientDetailActivity :
     }
 
     patientDetailConfig =
-      configurationRegistry.retrieveConfiguration<PatientDetailsViewConfiguration>(
+      configurationRegistry.retrieveConfiguration(
         configClassification = QuestConfigClassification.PATIENT_DETAILS_VIEW
       )
-
-    // TODO Load binary resources
-    profileConfig =
-      AssetUtil.decodeAsset(fileName = QuestPatientDetailViewModel.PROFILE_CONFIG, this)
 
     if (configurationRegistry.isAppIdInitialized()) {
       configureViews(patientDetailConfig)
     }
-    patientViewModel.run {
-      getDemographicsWithAdditionalData(patientId, patientDetailConfig)
-      getAllResults(patientId, profileConfig, patientDetailConfig)
-      getAllForms(profileConfig)
-    }
+
+    loadData()
+
     setContent { AppTheme { QuestPatientDetailScreen(patientViewModel) } }
   }
 
@@ -104,10 +97,19 @@ class QuestPatientDetailActivity :
   override fun onResume() {
     super.onResume()
 
+    loadData()
+  }
+
+  fun loadData() {
     patientViewModel.run {
       getDemographicsWithAdditionalData(patientId, patientDetailConfig)
-      getAllResults(patientId, profileConfig, patientDetailConfig)
-      getAllForms(profileConfig)
+      getAllResults(
+        patientId,
+        ResourceType.Patient,
+        patientDetailConfig.questionnaireFilter!!,
+        patientDetailConfig
+      )
+      getAllForms(patientDetailConfig.questionnaireFilter!!)
     }
   }
 
@@ -153,7 +155,7 @@ class QuestPatientDetailActivity :
             data?.getStringExtra(QUESTIONNAIRE_RESPONSE)?.let {
               val response =
                 FhirContext.forR4Cached().newJsonParser().parseResource(it) as QuestionnaireResponse
-              response.getEncounterId().let {
+              response.getEncounterId()?.let {
                 startActivity(
                   Intent(this, SimpleDetailsActivity::class.java).apply {
                     putExtra(RECORD_ID_ARG, it.replace("#", ""))
@@ -245,7 +247,7 @@ class QuestPatientDetailActivity :
       configClassification = QuestConfigClassification.RESULT_DETAILS_NAVIGATION
     )
 
-  override fun configureViews(viewConfiguration: PatientDetailsViewConfiguration) {
+  override fun configureViews(viewConfiguration: DataDetailsListViewConfiguration) {
     patientViewModel.updateViewConfigurations(viewConfiguration)
   }
 }