+++ android/quest/src/main/java/org/smartregister/fhircore/quest/ui/patient/register/PatientRegisterActivity.kt
@@ -34,7 +34,10 @@ import org.smartregister.fhircore.engine.util.extension.getDrawable
 import org.smartregister.fhircore.quest.R
 import org.smartregister.fhircore.quest.configuration.view.NavigationConfiguration
 import org.smartregister.fhircore.quest.configuration.view.NavigationOption
+import org.smartregister.fhircore.quest.configuration.view.QuestionnaireDataDetailsNavigationAction
 import org.smartregister.fhircore.quest.configuration.view.QuestionnaireNavigationAction
+import org.smartregister.fhircore.quest.ui.patient.details.QuestionnaireDataDetailActivity
+import org.smartregister.fhircore.quest.ui.patient.details.QuestionnaireDataDetailActivity.Companion.CLASSIFICATION_ARG
 import org.smartregister.fhircore.quest.ui.task.PatientTaskFragment
 import org.smartregister.fhircore.quest.util.QuestConfigClassification
 
@@ -84,17 +87,19 @@ class PatientRegisterActivity : BaseRegisterActivity() {
 
   override fun onNavigationOptionItemSelected(item: MenuItem): Boolean {
     when (item.itemId) {
-      R.id.menu_item_clients -> switchFragment(mainFragmentTag())
+      R.id.menu_item_clients -> switchFragment(mainFragmentTag(), isFilterVisible = false)
       R.id.menu_item_tasks ->
         switchFragment(
           tag = PatientTaskFragment.TAG,
           isRegisterFragment = false,
+          isFilterVisible = false,
           toolbarTitle = getString(R.string.tasks)
         )
       R.id.menu_item_settings ->
         switchFragment(
           tag = UserProfileFragment.TAG,
           isRegisterFragment = false,
+          isFilterVisible = false,
           toolbarTitle = getString(R.string.settings)
         )
       else ->
@@ -140,6 +145,12 @@ class PatientRegisterActivity : BaseRegisterActivity() {
                 formName = navigationOption.action.form,
               )
             )
+        )
+      is QuestionnaireDataDetailsNavigationAction ->
+        startActivity(
+          Intent(this, QuestionnaireDataDetailActivity::class.java).apply {
+            putExtra(CLASSIFICATION_ARG, navigationOption.action.classification)
+          }
         )
     }
   }